pyproject.toml
[project]
name = "maf-workflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agent-framework>=1.0.0b251120",
    "azure-core>=1.37.0",
    "lagom>=2.7.7",
    "openai>=2.11.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pip-audit>=2.10.0",
    "pre-commit>=4.5.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "pyright>=1.1.407",
    "ruff>=0.14.9",
]

[tool.ruff]
exclude = ["typings", ".venv"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "TID"]

[tool.ruff.format]
docstring-code-format = true

[tool.pyright]
venvPath = "."
venv = ".venv"
exclude = ["typings", ".venv", ".vscode", ".git"]

[tool.coverage.report]
fail_under = 90

[tool.coverage.run]
omit = [
    "*/__init__.py",
    "maf_workflow/protocols/*.py",
    "maf_workflow/hosting.py",
    "maf_workflow/agents/*.py",
    "maf_workflow/executors.py",
    "maf_workflow/workflow.py" 
]
